refactor: extract `getRow` to `getWeekDates`
